+++ chroma-common.spec
@@ -1,5 +1,5 @@
-%{!?name: %define name chroma-common}
-%{?!version: %define version %(%{__python} -c "from chroma_common import version; sys.stdout.write(version())")}
+%{!?name: %define name iml-common}
+%{?!version: %define version %(%{__python} -c "from iml_common import version; sys.stdout.write(version())")}
 %{?!package_release: %define package_release 1}
 %{?!python_sitelib: %define python_sitelib %(%{__python} -c "from distutils.sysconfig import get_python_lib; import sys; sys.stdout.write(get_python_lib())")}
 
@@ -36,7 +36,7 @@ rm -rf %{buildroot}
 mkdir -p $RPM_BUILD_ROOT/usr/sbin/
 
 touch test.files
-for test_file in $(find -L $RPM_BUILD_ROOT%{python_sitelib}/chroma_common/test/ -type f -name '*.py*'); do
+for test_file in $(find -L $RPM_BUILD_ROOT%{python_sitelib}/iml_common/test/ -type f -name '*.py*'); do
   install_file=${test_file/$RPM_BUILD_ROOT\///}
   echo "${install_file}" >> test.files
 done
@@ -45,7 +45,7 @@ done
 rm -rf %{buildroot}
 
 %files
-%exclude %{python_sitelib}/chroma_common/test
+%exclude %{python_sitelib}/iml_common/test
 %{python_sitelib}/*
 
 %files -f test.files test
