+++ tests/lib/test_sandbox_exception.py
@@ -1,7 +1,7 @@
 import mock
 
 from django.utils import unittest
-from chroma_common.lib.exception_sandbox import exceptionSandBox
+from iml_common.lib.exception_sandbox import exceptionSandBox
 
 
 class FencingTestCase(unittest.TestCase):
