+++ tests/lib/test_util.py
@@ -1,6 +1,6 @@
 import mock
 from django.utils import unittest
-from chroma_common.lib.util import PreserveFileAttributes
+from iml_common.lib.util import PreserveFileAttributes
 
 
 class PreservePermissionTestCase(unittest.TestCase):
